atoi returns a long

+++ ft_atoi.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: meltremb <meltremb@student.42quebec>       +#+  +:+       +#+        */
+/*   By: meltremb <meltremb@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/30 16:28:00 by meltremb          #+#    #+#             */
-/*   Updated: 2022/04/05 13:32:30 by meltremb         ###   ########.fr       */
+/*   Updated: 2023/04/05 14:54:16 by meltremb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-/* that fucking little fuck sucked probably 10 years of my life expectancy.
+
  * char--->int and the rules that look useless are for under min int and over
  * long max. You don't need them but i'm stupid. */
 
 #include"libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	long	nb;
 	long	i;
